requirements.txt
asgiref==3.8.1
Django==4.1.0
django-cors-headers==4.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
install==1.3.5
packaging==24.1
psycopg2-binary==2.8.4
PyJWT==2.8.0
sqlparse==0.5.0
typing-extensions==4.12.2
